Clean up load path and load node defaulting.
We don't want to use special load path values, just an empty string (we weren't using the constant except in one place for assignment).

For load node values we need an actual node so we don't run into API problems querying it. We could flag that node in the future to signify that it's a "built in".