Problems with inheriting a slot that we don't want
Fixed issue with regards to inheriting a slot via the standard prototype chain, so that it looks like it's present on the 'main type' (and is therefore considered 'most specific' and will always take precedence), but is not *own*ed by the 'main type'. This inserts checks to make sure that it is, and if it's not and the slot represents a method, that method's *owner* is used as a potential resolving source for the trait. This allows traits to work like how they should.