Fixed multiple issues with PouchDB update
Added the capability for the 'updateItem' action to also create the item if it couldn't find it (and renamed the action to be indicative of that). Also, fixed a misunderstanding around updating (needing to submit the whole document with changed fields) so that updates happen properly now. Also, make sure to convert the Date timestamp to a Number.