Change Attribute Node's global ID computation
We need to handle the path to the Element differently here to support 'TIBET barename syntax' extensions. It's not sufficient to just tack on the expression to the attribute to the fragment that is the ID of the element. We need to generate an XPath that includes the element instead.