Fix broken FunctionRequest test.
The fire() call is supposed to "reset" a signal. Which implied clearing any deferred. But that means if you fire() after setting up a then() you are out of sequence and won't see what you expect. Best to fire() the request and let it process either sync or async (start it running after all) and then invoke then() on the return value which will be the Response.