Changes to support proper output ordering
In order to get logging and other asynchronous output to be ordered properly, the shell was altered to generate an output cell before eval()ing. We leverage that here and made a variety of changes including having multiple coalescing records, recognizing new flags for logging output (including putting a special class on logging output that hides the input readout), and not outputting anything when trying to just construct a cell.