Don't overwrite key in handlers if it exists
Depending on how this call is called (with dontTraverse or not), we can end up with a different answer. If the key has already been computed, but we're just calling it with different semantics around dontTraverse, we don't just want to blindly blow away with TP.NO_RESULT. Otherwise, the handler will never be found again - the presence of TP.NO_RESULT will ensure that the computation is never done again.