Changed API so that 'ops' is no longer 'additive'
The 'ops' parameter to TP.$tokenize() (and, therefore, TP.$condenseJS()) now expects a complete set of operators. It no longer adds the standard set of JS operators to whatever is handed in. If 'ops' isn't supplied, this parameter will default to the standard set of JS operators (but only if 'ops' isn't supplied).