Enhance checks around whether to install resolver
We need to not only make sure that the slots themselves are not the same, but that they also don't point to the same resolution functions before we decide to install a resolver. This is the same check done in the addTraits() call, but is necessary here due to inherited properties that might be resolved properties themselves.