Fix major issue with redrawing content
If we don't clone this node content, then the original DOM is modified. This is bad because the script nodes are removed as part of the drawing process. This causes subsequent rendering of the content to not have any script nodes and for the hook file, etc. to not activated because those references are gone.