Updated logic for responder checking for Nodes
Now, in addition to calling an 'isResponderFor'+signal type method for a Node, this routine will try to find out if the receiving node defined a handler for that signal and, if it did, it will return true. Note that we had to remove the 'isResponderForSignal' top-level backstop or otherwise the 'handler check' never got called.