Install a spy on TP.raise() to track raises
We need to be able to track TP.raise() calls (the base, core call) so that we can know if we need to fail the test. Other calls in TIBET will try...catch around Errors thrown from TP.raise() such that they will never make it 'up' to the test harness. This is by design, but the harness needs to know when these calls are made, hence the spy.