Changed calculation of GID for Document
Document GIDs are now computed from the window.location if a cached value doesn't already exist in the TP.SRC_LOCATION slot on the document itself (which is also a change - it used to be stored on the document element). Also, TP.SRC_LOCATION is now stamped onto the document *before* processing, so that when the GID is computed (and cached) during load event processing, we have a more complete value than just using the xml:base or base.