No need to create a TP.core.Hash here
We don't need to create a TP.core.Hash instance each time here - just access the instance-side prototype, which is the same thing as what we were trying to obtain here. Also, update the comments to remove the text around converting Dates and RegExps when 'smart converting' - we don't do that anymore.