Add app root parameter to PhantomJS-invoked URLs.
Phantom can't determine app root properly since it's activated from a library path which can't know whether a project uses tibet_pub or whether it's in the TIBET library etc. We repair that by pushing app root from package data in the CLI and then parsing in the loader for a URL-based hint.