Only signal Change from the primary URI
We no longer signal Change from each affected secondary URI, but only from the primary URI. We do keep a list of the secondary URIs that changed and supply that in the signal. Also, updated the terminology from 'sub' to 'secondary' to be consistent with the same change made elsewhere in this type.