Update tag command to be more atomic and include tests.
Needed to avoid a "copy then rename" approach so this version renames as it writes each file. The other main change for watcher compatibility is that the config file update is done _last_ after all other processing is complete and the files are written. This update also adds an entry to the app_tests target in standard package files for a tag test file.