Support failing of tests that log errors
Updated the log entry code and test harness code to support a special 'log appender' that will fail a currently executing test case if, in the process of running the test case, an error (or worse) is logged. Also, this is put behind a cfg flag so that users can switch if off if they so desire.