Various updates to make tests more robust
Made naming consistent so that for tests with multiple people, we use the '...people' URI. Also, removed hardcoded IDs and their corresponding ID refs in the 'repeat' tests and use numeric indexing to get the field to test. Also, adjusted insert/delete signaling to supply the origin and not rely on target. Lastly, for now, we're skipping the 'repeat paging' tests.