Give the harness control over resolving the case
The test case itself shouldn't be controlling the Promise that represents the 'whole case'. In fact, this causes issues with the internal Promise when the case adds more Promises onto the internal Promise. Instead, we manually call the resolver of the 'whole case' when its appropriate that all other Promises for the case itself (not the afterEach or after, those come later) have resolved.