Updated usage of TP.isURI() to be a type check
This method currently takes both Strings and TP.core.URI instances, which does not follow TIBET coding standard (i.e. is* methods are typically type checks). This is a initial cut at conversion so that TP.isURIString() should be used if the data to be tested is a URI string, not a TP.core.URI instance. A warning has been added to TP.isURI() for when a String is supplied to that method.