Cleaned up and made more adjustments for skips
Skipped tests can not only come as 'whole suite' skips that can't be tracked by look for '# SKIP', but also as a 'suite that has an only() case and therefore treats all of the others as skipped'. This patch fixes that case as well and cleans up a bit of the logic around 'info' vs. 'result'.