Update attribute mgmt to adapt to XML semantics
Recent browser changes mean that the XML rules around attributes and namespaces are now fully implemented. This means that attribute not specifically prefixed are in the 'null' namespace. We provide a convenience, such that if a prefix was supplied to these routines such that its corresponding namespace matches the namespace of the element, we go ahead and perform the action unprefixed.
Note that we do not update 'elementSetAttribute()' here, as a convenience mechanism has a number of unintended consequences.