Adjusted to avoid fail()ing tests due to raise()
The checks for whether a TP.raise() occurred (which it now will as part of an HTTP request failing) occurs after each invocation of afterEach(). Move disabling the debugger and setting the log level in anticipation that there will be failures here (and, in fact, we're expecting them) into the 'whole suite' before() and after() methods. Also add the disabling/enabling suspending of whether we care about TP.raise()s.