Repair glitch around exception firing matching incorrectly on spoofs.
We don't want the iterations of exception handling to trigger a "more specific" handler. Prior to this patch you could fire a 'FileAccessException' and the handler would be invoked for both that and 'Exception', the latter invocation is incorrect since it's not a handler for that "generic level" of exception.