Return null from getResourceURI() on XHTML elements
This method is not necessary on XHTML elements and, in fact, is misleading. Also, note how we manually resolve the trait on the xctrls:codeeditor because it mixes in TP.html.textUtilities, subtyped from TP.html.Element and it picks up this version of this method, when it really wants to stick with its mainline version.