Added flag and mechanism to define undefined prefixes
Sometimes, especially during the authoring process, the XML parser will be asked to parse tags where the prefix hasn't been defined. When this flag is on, when the XML parser throws an error, TIBET will attempt to 'autodefine' the prefix (using 'urn:temp') and then parse the document again (leveraging the 'auto namespacing' functionality).