Rewrote TP.elementGetOffsetParent() to match spec
This routine is now spec compliant (per CSS OM), will test first to see if there is an '.offsetParent' and return that and, if not, call into a lower-level routine. Also, the E5 getter for '.offsetParent' used in the boot code has been updated to use the lower-level routine to avoid endless recursion.