Made the Karma bridge use console.log for Errors
If we report test Errors (i.e. where a test has actually 'throw'n) via the normal Karma logging of a result, Karma will count the test twice since we log twice. We filter those here and just log the Error output with a regular 'console.log()' invocation (which Karma will output to the Terminal as a regular log message, not 'counting' it).