Changed ‘TP’ versions of alert, confirm & prompt to return Promises
Updated any usages of these that we want to keep to expect a returned Promise. Also, removed a lot of usages of these in examples, in favor of ‘TP.info()’ instead. Lastly, any really old code that uses this functionality got changed back to using the native versions since that code depends on ‘blocking’ behavior and the TP.* versions are no longer blocking.