Make the ‘handled’ slot observer ID specific
Since we now have multiple observers coming through this code, we need to be specific about *which* observer ‘handled’ it and allow each to handle it.
Also, we now stash the registry record on the observer object itself, which gets handed to the callback function as a parameter, thereby making things a bit cleaner.