Pass along the new node in the callback Function
This pertains to primitives or wrappers that insert, set, replace, etc. content on HTML, XML, documents or elements. We want the new node back that got placed into the document (which is not always the same as the one that got handed to the routine, because of node importing, etc.)