Changed ignoreHandler(), etc to take context object
It’s important to take into account the context object of the executing handler to determine whether it should be ignored or not. Handler functions (i.e. methods) can be shared amongst different objects via inheritance, etc. and need to be uniqued against the object as well as the handler.