Removed native platform calls to compute applied rules
The few platforms that supported the ability to get the rules that applied to an element are removing those calls, so this commit removes that platform-level code and updates the ‘TIBET way’ of obtaining the same results. Also, we now use a proper constant (TP.APPLIED_RULES) to store the per element caches (for consistency with the rest of the platform).