Completely restructured to use data-based selection model
Abstracted getting/setting the selection model so that types that mix in TP.core.SelectingUIElementNode are no longer assumed to have ‘selected elements’ and ‘value elements’. ‘selected elements’ went away and ‘value elements’ and all of that attendant machinery that drives the selection in those cases moved to TP.core.TogglingUIElementNode. Also, the ‘render()’ method concept was introduced to allow subtypes a hook to override rendering the selection process.