Add support for ‘opaque bubbling’ for DOM signals
We have support for ‘opaque capturing’ when computing the resolved target. We need analogous support for ‘opaque bubbling’ (i.e. won’t bubble beyond a certain parent) in the DOM_FIRING policy. Note that this doesn’t affect most keyboard/mouse handling because that tends to be done with direct observer handler firing to a specific target (including state machines).