Restructured and fixed bugs to be much more efficient
Repurposed old parameters that don’t get used anymore, fixed issues with calling only the ‘update’ Function (i.e. must update the individual datums) and, lastly, not reallocating the ‘scrollerFunc’ each time ‘render()’ is called. This makes this object much faster and more intelligent about when it actually *needs* to re-render and, if so, if it should be calling the update() or the enter()/exit() selections.