Rename traverseUsing() to populateBayUsing() for clarity and refactored
Factored out some of the final updating functionality into it’s own method and called it separately after all of the traversal is done. Also, we now throw the InspectorDidFocus signal at the end of traversePath(), which is why we don’t need to do it separately where we use traversePath().