Reworked to not require a trigger element
The trigger element is only really useful to position the popup. If the trigger point is supplied anyway, there’s no need to supply or use the trigger element. This is important when computing whether to dismiss the popup.
Also, adjustments were made to fetch a synchronous URI’s content directly.