Restore command to build our app
Now that projects use the PhantomJS that’s built into TIBET, this command doesn’t require the user to install PhantomJS into their project. Also, if this command isn’t run, then the ‘bare’ URL (i.e. production package) will error when loading, so we want to encourage the developer to run this.