Filter out more types of DOM mutations
If the node that changed is a CDATA Section or a comment node, we want to use the same logic as when it’s a text node. Also, if the element has a ‘type’ attribute that is the same as TP.ietf.Mime.TIBET_CSS, we don’t want to process it (it’s a ‘tibet:style’ element). Lastly, we also filter out setting the ‘tibet:originalHref’ attribute.