Fix longstanding bug with using a spoofed signal name when dealing with mulitple signal names
If an Array of signal names was supplied to these methods, any of the signal names that were spoofed were turned into undefined values because they didn’t have a concrete type associated with them. This is different than the logic that deals with singular signal name values, which would just use the spoofed name if a concrete type couldn’t be found.