No longer cache the MIME type for the URI
Since the MIME type can change as the endpoint wants it to, it doesn’t make sense to cache the MIME type. Also, there are times that we want to compute the MIME type with a new resource rather than the existing resource, so we now have an optional parameter for that.