Always send HaloDidFocus/HaloDidBlur or otherwise other parts of the HUD will not update properly
When an element is ‘recast’, even though *it* is the same element, subelements of it are not. Other parts of the HUD, like sidebar elements that observe the halo target, still need to be told to refresh themselves so that they don’t have stale references to old parts of the DOM that are gone.