Changed activate/deactivate/remove of managed MutationObservers so that a null target can be supplied
Now, the activate/deactivate methods will only manage the underlying browser’s MO machinery to add/remove the native MO. This now leaves the targetNode and native MO of a particular entry in place when deactivated so that it can be reactivated simply by providing the name and it will reuse the prior target node. This also avoids creating a native MO each time the managed MO is activated. Lastly, the ‘active’ state of the MO is tracked.