Only update UI aspects that match the model aspect being changed.
This is a big perf win in cases where a binding expression has multiple UI aspects it is managing. It also allows more intelligent usage of aspects that shouldn’t be updated (or are one-shot, like defaultValue) and would therefore cause issues if they were updated because another aspect defined in the same binding attribute was changed.