Fixed issue with URL attributes getting rewritten even though they didn’t contain URLs
In SVG, some attributes can have both URL and non-URL values (i.e. ‘fill’, etc), unlike in (X)HTML, where an attribute is either a ‘URI’ attribute or not. This is disambiguated by the requirement in SVG to use the ‘url(…)’ syntax in the attribute value. By overriding the tagResolve method and providing this additional logic, we avoid having attributes in SVG rewritten as URLs when that’s not appropriate.