Abstract resource comparison into an overrideable method to test for dirtiness
Rather than allowing the caller to set a flag to track whether a URI is dirty, which is now removed, this mechanism provides an overrideable method to do resource comparison for the purpose of tracking dirtiness. Note that the top-level TP.core.URI type compares by equality, but that the TP.core.URN type overrides that and compares by identity.