Remove the main PouchDB module from TIBET and substitute a browser-only PouchDB version
We do this to avoid installing LevelDB, etc. which is required to run PouchDB on the server, which we don’t use (and bloats the TIBET production package by requiring C build tools, etc).
Unfortunately, building the browser-only version of PouchDB is not straightforward. See the updated _rollup_pouchdb.js file (basically, we have to do it manually).