Keep track of the refreshed elements keyed by aspect and facet
Otherwise, we might have refreshed one particular aspect or facet of an element and consider that to be ‘updated’, which isn’t true if we need to update another aspect or facet.
Note that we need to flatten/unique the list before sending the native signal for API compatibility with existing code.