If the component type isn’t defined, make sure to import the component definition scripts
This will occur if React/ReactDOM is already loaded and we’re putting new components into the page dynamically. Sometimes a component is loaded along with React/ReactDOM as the ‘first load’, but sometimes it is loaded long after React/ReactDOM has been loaded and it needs to be defined.