Set a flag and don’t execute the following steps in the Promise if the content is the same.
This is due to this code being ported over to being a Promise. The `return` statement used to return from the whole callback (when this code was a callback), but now we need to set a flag and check it for the rest of the Promise chain (because it’s not possible terminate a Promise chain ‘early’).