Do not replace the lib root references - this causes multiple problems after thawing
Upon further examination, this is unnecessary and causes issues when thawing. The reason is that the ‘freeze’ command changes paths from ‘node_modules/tibet’ to ‘TIBET-INF/tibet’ when this is unnecessary (there are no app-level files that have embedded references to ‘node_modules/tibet’). But there are other references to ‘TIBET-INF/tibet’. So, when the app is thawed and the ‘thaw’ command wants to put all references to ‘TIBET-INF/tibet’ back to ‘node_modules/tibet’, it actually corrupts the original references to ‘TIBET-INF/tibet’ and the app stops working.