Use a ‘proxified method’ technique to handle shadowed methods on TP.core.Hash
If the slot exists both on the TP.core.Hash object and the internal $$hash of that object, then we need to do some further trickery around supporting both ‘get access’ (i.e. for things like destructuring a TP.core.Hash) and for method invocation on that hash itself (i.e. if you have an ‘at’ property in the internal hash, how do you disambiguate?). The only way to do this is to install a Function Proxy as the ‘method’ on the hash once a property with the same name has been added.