Rework TP.core.Worker to be more Promise-based and remove cruft
This entailed a variety of changes including removing the ‘PromiseWorker’ type and moving its Promise-based invocation method ‘up’ to TP.core.Worker (leveraging it in all calls) and renaming the class to be ‘generic worker’ (since it uses our ‘generic’ worker script). Also, the worker script’s ‘eval’ method now returns it’s value and there is a ‘evalNoReturn’ method which is invoked when the caller on the main thread does *not* want the return value passed back (like when it’s a Function - which breaks the serialization algorithm).