When computing a possible source type, we don’t want to check already resolved method values
Unless the slot itself is the same, here we always continue until we find a real sourceType. We very well might run into a situation where, between trait types and our supertype (somewhere in the hierarchy) the ‘computed resolved’ values are the same, but we don’t care. We want to keep going until we have computed a trait type.