Reworked setDisplayValue to ask the list directly for the label matching the value.
This needs to be done this way since the list will probably be closed (and most of its items gone due to the resizing that happens) and so for items after the 2nd or 3rd ones, the ‘itemWithValue’ will not point to anything meaningful. Therefore, we ask the list directly, which will use its model data directly.