Another round of updates to the ECMAScript Module loader
We now no longer use our scheme only for file:// loading. Since we need to support 'bare specifier' module names and the ECMAScript module loader is still a work in progress (and insufficient for our needs) we need a way to map specifiers and so we use the TIBET cfg XML to do so.

Also, we no longer use the silly convention of local file names for loading modules in a packaged environment using SystemJS - we now use proper 'full path' (well, full virtual path).